src:core:model: Add operator =() for Ptr when assigning naked pointer
Added operator =(T *) for Ptr class, so when assigning the naked pointer,
we wouldn't take the additional reference on it.

This is based on fact, that naked pointer isn't managed by Ptr class,
and nobody isn't going to take the additional reference from it which
will create the  memory leak.